+++ lib/src/analyzers/unused_l10n_analyzer/models/unused_l10n_issue.dart
@@ -1,0 +1,21 @@
+import 'package:meta/meta.dart';
+import 'package:source_span/source_span.dart';
+
+/// Represents an issue detected by the unused localization check.
+@immutable
+class UnusedL10nIssue {
+  /// The name of a class member, which is unused
+  final String memberName;
+
+  /// The source location associated with this issue.
+  final SourceLocation location;
+
+  /// Initialize a newly created [UnusedL10nIssue].
+  ///
+  /// The issue is associated with the given [location]. Used for
+  /// creating an unused localization report.
+  const UnusedL10nIssue({
+    required this.memberName,
+    required this.location,
+  });
+}