+++ test/resources/unused_l10n_analyzer/test_i18n.dart
@@ -1,0 +1,7 @@
+class TestI18n {
+  static const String field = 'field';
+
+  static String get getter => 'getter'; // LINT
+
+  static String method(String value) => value;
+}