+++ test/resources/unused_l10n_analyzer/unused_l10n_analyzer_example.dart
@@ -1,0 +1,9 @@
+import 'test_i18n.dart';
+
+void main() {
+  final _ = TestI18n.method('value');
+}
+
+class SomeClass {
+  final field = TestI18n.field;
+}